Rename files. Remove import aliases
